pyproject.toml
[tool.poetry]
name = "horoscopebot"
description = "Tells lies"
version = "1.0.0"
authors = ["OpenAlcoholics <team@openalcoholics.group>"]
license = "MIT"

[[tool.poetry.source]]
name = "internal"
url = "https://pypi.bjoernpetersen.net"
secondary = true

[tool.poetry.dependencies]
certifi = "2022.6.15.1"
Deprecated = "^1.2.13"
google-cloud-pubsub = "^2.0.0"
openai = "^0.23.0"
pendulum = "^2.0.0"
python = "~3.10"
python-dotenv = "^0.21.0"
rate_limit = { version = "1.0.3", source = "internal" }
requests = "2.28.1"
sentry-sdk = "1.9.8"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^5.0.0"
mypy = "0.971"
pytest = "^7.1.2"
types-Deprecated = "^1.2.9"
types-requests = "2.28.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true
